test: run the writer-resolution parity pin in the Integration suite (#105 amendment)
The parity check resolves Verdict's own EvidenceWriter binding, which only
the Integration harness boots; the Feature posture suite stays independent
of Verdict's container.
